==> 01_FeatureExtractionPart01.ipynb <==
## Initial Setup

# !pip install pyspark

from pyspark.sql import SparkSession

spark = SparkSession.builder.appName("Basics").getOrCreate()

spark

# mounting drive
# from google.colab import drive
# drive.mount('/content/drive')

from pyspark.sql.types import StructType, IntegerType, StringType, FloatType,ArrayType
from pyspark.sql.functions import split, col, regexp_replace , round as round2 , coalesce, lit
import pyspark.sql.functions as F


## Read the data

raw_recipes_df = spark.read.csv('s3a://raw-recipes-clean-upgrad/RAW_recipes_cleaned.csv', 
                                header= True,
                                inferSchema= True
                               )


raw_recipes_df.show(5)

# Code check cell
assert raw_recipes_df.count() == 231637, "There is a mistake in reading the data."
assert len(raw_recipes_df.columns) == 12, "There is a mistake in reading the data."

## Extract ```nutrition``` values 

raw_recipes_df.select('nutrition').show(5, truncate=False)

# List of nutrition columns

nutrition_column_names = ['calories',
                          'total_fat_PDV',
                          'sugar_PDV',
                          'sodium_PDV',
                          'protein_PDV',
                          'saturated_fat_PDV',
                          'carbohydrates_PDV']

# STEP 1 - string operations to convert nutrition to an array 

raw_recipes_df = raw_recipes_df. withColumn('nutrition',regexp_replace(regexp_replace('nutrition','\[',''),'\]','') )
                
raw_recipes_df.show(5)

raw_recipes_df2 = raw_recipes_df

# STEP 2 - create new columns with individual nutrition values 

raw_recipes_df2.createOrReplaceTempView('Recipes')

raw_recipes_df= spark.sql("SELECT *, cast(replace(split(nutrition,',')[0],'[','') as float) as calories,"
                " cast(split(nutrition,',')[1] as float) as total_fat_PDV,"
                " cast(split(nutrition,',')[2] as float) as sugar_PDV,"
                " cast(split(nutrition,',')[3] as float) as sodium_PDV,"
                " cast(split(nutrition,',')[4] as float) as protein_PDV,"
                " cast(split(nutrition,',')[5] as float) as saturated_fat_PDV,"
                " cast(replace(split(nutrition,',')[6],']','') as float) as carbohydrates_PDV"
                " from Recipes;")

# Code check cell
assert raw_recipes_df.schema["carbohydrates_PDV"].dataType == FloatType(), "Recheck your typecasting"
assert raw_recipes_df.collect()[123432][14] == 62.0, "The columns have not been split correctly."
assert raw_recipes_df.collect()[10000][12] == 60.400001525878906, "The columns have not been split correctly."

## Make nutrition-per-100 calorie columns

By converting the nutrition values from absolute to relative terms, we ensure that portion size is not a factor in the analysis. 


Naming convention: Original column name ```total fat (PDV)```, column name after column ```total_fat_per_100_cal```

for nutrition_col in nutrition_column_names:
    if nutrition_col != 'calories':
        nutrition_per_100_cal_col = (nutrition_col
                                 .replace('_PDV','')
                                 +'_per_100_cal')
        raw_recipes_df = raw_recipes_df.withColumn(nutrition_per_100_cal_col, coalesce(round2(raw_recipes_df[nutrition_col]/raw_recipes_df['calories']*100,2),lit(0)).cast('float'))                         
        

raw_recipes_df.show(5)

# total fat check for id 28881
assert raw_recipes_df.filter("id == 28881").select('total_fat_per_100_cal').first()[0] == 0, "total_fat_per_100_cal for recipe 28881 should be 0"

# total fat check for id 112140
assert round(raw_recipes_df.filter("id == 112140").select('total_fat_per_100_cal').first()[0]) == 8, "total_fat_per_100_cal for recipe 112140 should be 8"

## Convert `Tags` to `array` type

raw_recipes_df.select('tags').show(5, truncate=False)

raw_recipes_df = raw_recipes_df \
                  .withColumn('tags', \
                  split( \
                    regexp_replace( \
                      regexp_replace( \
                        regexp_replace( \
                          regexp_replace( \
                            raw_recipes_df['tags'], \
                            '\[',''), \
                            '\]',''), \
                            '\'',''), \
                            ' ',''), \
                            ','))

raw_recipes_df.select('tags').show(5, truncate=False)

# Code check cell
assert raw_recipes_df.schema["tags"].dataType == ArrayType(StringType(), True), "You have not split the string into an array."
assert raw_recipes_df.collect()[2][5] == ['time-to-make','course', 'preparation', 'main-dish', 'chili', 'crock-pot-slow-cooker', 'dietary', 'equipment', '4-hours-or-less'], "Recheck your string cleaning and splitting operations."

## Join Recipe Data to Review Data

# Reading the second data set. 
raw_ratings_df = (spark.read.csv("s3a://raw-interactions-upgrad/RAW_interactions_cleaned.csv",header=True,inferSchema=True ).withColumn("review_date",  F.col("date")).drop(F.col("date")))

raw_ratings_df.printSchema()

# Code check cell
assert raw_ratings_df.count() == 1132367, "There is a mistake in reading the data."
assert len(raw_ratings_df.columns) == 5, "There is a mistake in reading the data."

raw_ratings_df.show(5)

interaction_level_df = raw_ratings_df.join(raw_recipes_df,raw_ratings_df.recipe_id == raw_recipes_df.id,'left')

# Code check cell
list1 = raw_ratings_df.select('recipe_id').collect()
list2 = raw_recipes_df.select('id').collect()
exclusive_set = set(list1)-set(list2)

assert len(exclusive_set) == 0, "There is a mistake in reading one of the two data files."

interaction_level_df.show(5)

## Recipe Submission Date Based Features

**Steps:**
1. Convert ```submitted``` column to `DateType()`
2. Use review date to derive columns `days_since_submission_on_review_date`,  `months_since_submission_on_review_date`,  `years_since_submission_on_review_date`

interaction_level_df = (interaction_level_df
                        .withColumn('submitted', F.to_date(interaction_level_df['submitted']) 
                                   ))

interaction_level_df = (interaction_level_df \
                        .withColumn('days_since_submission_on_review_date', \
                                   F.datediff(interaction_level_df['review_date'],interaction_level_df['submitted']) \
                                   ) \
                        .withColumn('months_since_submission_on_review_date', \
                                   F.months_between(interaction_level_df['review_date'],interaction_level_df['submitted']) \
                                   ) \
                        .withColumn('years_since_submission_on_review_date', \
                                   F.months_between(interaction_level_df['review_date'],interaction_level_df['submitted'])/lit(12) \
                                   ))

interaction_level_df.show(5)

## Save the data we have created so far in a parquet file. 

interaction_level_df.printSchema()

## Write the raw_recipes_df
## Write data to the S3 bucket which will be your working directory 

interaction_level_df.write.mode('overwrite').parquet(    's3://recass/interaction_level_df.parquet'   )
