==> 03_FeatureExtractionPart02.ipynb <==
## Initial Setup

# !pip install pyspark

from pyspark.sql import SparkSession

spark = SparkSession.builder.appName("Basics").getOrCreate()

spark

# Run this everytime you create a new spark instance. 

spark.sparkContext.install_pypi_package("plotly==5.5.0")
spark.sparkContext.install_pypi_package("pandas==0.25.1")
spark.sparkContext.install_pypi_package("numpy==1.14.5")
spark.sparkContext.install_pypi_package("matplotlib==3.1.1")

from pyspark.sql.types import StructType, IntegerType, StringType, FloatType,ArrayType
from pyspark.sql.functions import split, col, regexp_replace , round as round2 , coalesce, lit
import pyspark.sql.functions as F
from pyspark.sql.window import Window

# mounting drive
# from google.colab import drive
# drive.mount('/content/drive')

## Defining Custom Functions 

def get_quantiles(df, col_name, quantiles_list = [0.01, 0.25, 0.5, 0.75, 0.99]):
    """
    Takes a numerical column and returns column values at requested quantiles

    Inputs 
    Argument 1: Dataframe
    Argument 2: Name of the column
    Argument 3: A list of quantiles you want to find. Default value [0.01, 0.25, 0.5, 0.75, 0.99]

    Output 
    Returns a dictionary with quantiles as keys and column quantile values as values 
    """
    quantiles_list.append(0)
    quantiles_list.append(1)
    quantiles_list = list(set(quantiles_list))
    quantiles_dict = {}
    # Store min, quantiles and max in output dict, sequentially
    for qnt in quantiles_list:
      quantiles_dict[qnt] = df.select(F.percentile_approx(col_name,qnt,1000000)).collect()[0][0]

    return(quantiles_dict)

def add_OHE_columns (df, n_name_list):
    """
    Given a list of tags, creates one hot encoded columns for each tag. 
  
    Input
    Argument 1: Dataframe in which the function will add the new columns
    Argument 2: list of tags
  
    Output
    Prints the names of columns that have been added 
    Returns the modified dataframe 
    """
    for name in n_name_list:
        df = df.withColumn('has_tag_'+name, F.when(F.array_contains(col('tags'),name)== 'true', 1).otherwise(0))
        print ("added column: has_tag_"+name)

    return df

## Read the data

interaction_level_df = spark.read.parquet(
's3://recass/interaction_level_df_post_eda.parquet'
)

## Adding user level average features

1. user_avg_rating
2. user_n_ratings
3. user_avg_years_betwn_review_and_submission
4. user_avg_prep_time_recipes_reviewed
5. user_avg_n_steps_recipes_reviewed
6. user_avg_n_ingredients_recipes_reviewed

windowSpec = Window.partitionBy('user_id')
interaction_level_df = interaction_level_df.withColumn('user_avg_rating',F.avg(col('rating')).over(windowSpec))\
.withColumn('user_n_ratings',F.count(col('rating')).over(windowSpec))\
.withColumn('user_avg_years_betwn_review_and_submission',F.avg(col('years_since_submission_on_review_date')).over(windowSpec))\
.withColumn('user_avg_prep_time_recipes_reviewed',F.avg(col('minutes')).over(windowSpec))\
.withColumn('user_avg_n_steps_recipes_reviewed',F.avg(col('n_steps')).over(windowSpec))\
.withColumn('user_avg_n_ingredients_recipes_reviewed',F.avg(col('n_ingredients')).over(windowSpec)) 


Add columns 

7. user_avg_calories

8. user_avg_total_fat_per_100_cal_recipes_reviewed

9. user_avg_sugar_per_100_cal_recipes_reviewed

10. user_avg_sodium_per_100_cal_recipes_reviewed

11. user_avg_protein_per_100_cal_recipes_reviewed

12. user_avg_saturated_fat_per_100_cal_recipes_reviewed

13. user_avg_carbohydrates_per_100_cal_recipes_reviewed

interaction_level_df = interaction_level_df.withColumn('user_avg_calories',F.avg(col('calories')).over(windowSpec))\
.withColumn('user_avg_total_fat_per_100_cal_recipes_reviewed',F.avg(col('total_fat_per_100_cal')).over(windowSpec)) \
.withColumn('user_avg_sugar_per_100_cal_recipes_reviewed',F.avg(col('sugar_per_100_cal')).over(windowSpec)) \
.withColumn('user_avg_sodium_per_100_cal_recipes_reviewed',F.avg(col('sodium_per_100_cal')).over(windowSpec)) \
.withColumn('user_avg_protein_per_100_cal_recipes_reviewed',F.avg(col('protein_per_100_cal')).over(windowSpec)) \
.withColumn('user_avg_saturated_fat_per_100_cal_recipes_reviewed',F.avg(col('saturated_fat_per_100_cal')).over(windowSpec)) \
.withColumn('user_avg_carbohydrates_per_100_cal_recipes_reviewed',F.avg(col('carbohydrates_per_100_cal')).over(windowSpec))

# Code check cell
assert(round(interaction_level_df.filter('user_id == 601529').select('user_avg_rating').first()[0], 2) == 4.22)
assert(interaction_level_df.filter('user_id == 601529').select('user_n_ratings').first()[0] == 27)
assert(round(interaction_level_df.filter('user_id == 601529').select('user_avg_years_betwn_review_and_submission').first()[0], 2) == 3.51)
assert(interaction_level_df.filter('user_id == 233044').select('user_avg_prep_time_recipes_reviewed').first()[0] == 50.3)
assert(interaction_level_df.filter('user_id == 233044').select('user_avg_n_steps_recipes_reviewed').first()[0] == 8.8)
assert(interaction_level_df.filter('user_id == 233044').select('user_avg_n_ingredients_recipes_reviewed').first()[0] == 8.2)
assert(round(interaction_level_df.filter('user_id == 233044').select('user_avg_total_fat_per_100_cal_recipes_reviewed').first()[0]) == 6)

**More Features:**

high_ratings = 5 rating

- `user_avg_years_betwn_review_and_submission_high_ratings`
- `user_avg_prep_time_recipes_reviewed_high_ratings`
- `user_avg_n_steps_recipes_reviewed_high_ratings`
- `user_avg_n_ingredients_recipes_reviewed_high_ratings`

# Create indicator columns for highest ratings 
interaction_level_df.createOrReplaceTempView('interaction_level_df2')
interaction_level_df2 = spark.sql("SELECT user_id,recipe_id, rating, review, review_date, name, id, minutes, contributor_id, submitted, \
tags, nutrition, n_steps, steps, description, ingredients, n_ingredients, calories, total_fat_PDV, sugar_PDV, sodium_PDV, \
protein_PDV, saturated_fat_PDV, carbohydrates_PDV, total_fat_per_100_cal, sugar_per_100_cal, sodium_per_100_cal, protein_per_100_cal, \
saturated_fat_per_100_cal, carbohydrates_per_100_cal, days_since_submission_on_review_date, months_since_submission_on_review_date, \
years_since_submission_on_review_date, years_since_submission_on_review_date_bucket, prep_time_bucket, n_steps_bucket, \
n_ingredients_bucket, calories_bucket, total_fat_PDV_bucket, sugar_PDV_bucket, sodium_PDV_bucket, protein_PDV_bucket, saturated_fat_PDV_bucket, \
carbohydrates_PDV_bucket, total_fat_per_100_cal_bucket, sugar_per_100_cal_bucket, sodium_per_100_cal_bucket, protein_per_100_cal_bucket, \
saturated_fat_per_100_cal_bucket, carbohydrates_per_100_cal_bucket, user_avg_rating, user_n_ratings, user_avg_years_betwn_review_and_submission, \
user_avg_prep_time_recipes_reviewed, user_avg_n_steps_recipes_reviewed, user_avg_n_ingredients_recipes_reviewed, user_avg_calories, \
user_avg_total_fat_per_100_cal_recipes_reviewed, user_avg_sugar_per_100_cal_recipes_reviewed, user_avg_sodium_per_100_cal_recipes_reviewed, \
user_avg_protein_per_100_cal_recipes_reviewed, user_avg_saturated_fat_per_100_cal_recipes_reviewed, user_avg_carbohydrates_per_100_cal_recipes_reviewed \
FROM interaction_level_df2 ; " )
interaction_level_df2.createOrReplaceTempView('interaction_level_df3')

interaction_level_df = spark.sql("SELECT *,\
case when rating =5 then 1 else 0 end as highest_rating_flag, \
sum(case when rating =5 then 1 else 0 end  ) over(partition by user_id) as user_n_5_ratings,\
avg(case when rating =5 then years_since_submission_on_review_date end ) over(partition by user_id) as user_avg_years_betwn_review_and_submission_5_ratings, \
avg(case when rating =5 then minutes end ) over(partition by user_id) as user_avg_prep_time_recipes_reviewed_5_ratings, \
avg(case when rating =5 then n_steps end ) over(partition by user_id) as user_avg_n_steps_recipes_reviewed_5_ratings, \
avg(case when rating =5 then n_ingredients end ) over(partition by user_id) as user_avg_n_ingredients_recipes_reviewed_5_ratings, \
avg(case when rating =5 then calories end ) over(partition by user_id) as  user_avg_calories_recipes_reviewed_5_ratings,\
avg(case when rating =5 then sugar_per_100_cal end ) over(partition by user_id) as  user_avg_sodium_per_100_cal_recipes_reviewed_5_ratings,\
avg(case when rating =5 then protein_per_100_cal end ) over(partition by user_id) as  user_avg_protein_per_100_cal_recipes_reviewed_5_ratings,\
avg(case when rating =5 then saturated_fat_per_100_cal end ) over(partition by user_id) as  user_avg_saturated_fat_per_100_cal_recipes_reviewed_5_ratings,\
avg(case when rating =5 then total_fat_per_100_cal end ) over(partition by user_id) as  user_avg_total_fat_per_100_cal_recipes_reviewed_5_ratings,\
avg(case when rating =5 then carbohydrates_per_100_cal end ) over(partition by user_id) as  user_avg_carbohydrates_per_100_cal_recipes_reviewed_5_ratings  FROM interaction_level_df3;")

# leaving null values as it is , because it is an indicator of no rating being given as 5 stars

assert(interaction_level_df.filter('user_id == 233044').select('user_n_5_ratings').first()[0] == 7)
assert(round(interaction_level_df.filter('user_id == 233044').select('user_avg_years_betwn_review_and_submission_5_ratings').first()[0], 2) == 2.24)
assert(round(interaction_level_df.filter('user_id == 233044').select('user_avg_prep_time_recipes_reviewed_5_ratings').first()[0]) == 46)
assert(round(interaction_level_df.filter('user_id == 233044').select('user_avg_n_steps_recipes_reviewed_5_ratings').first()[0], 2) == 7.29)
assert(round(interaction_level_df.filter('user_id == 233044').select('user_avg_n_ingredients_recipes_reviewed_5_ratings').first()[0], 2) == 6.86)

interaction_level_df.printSchema()

## Tags level EDA

interaction_tag_level_df = interaction_level_df.withColumn('individual_tag',F.explode('tags'))

tags_ratings_summary = (interaction_tag_level_df
                        .groupBy('individual_tag').agg(F.avg('rating').alias('avg_user_rating'),
#                                                      F.max('rating').alias('max_user_rating'),
#                                                      F.min('rating').alias('min_user_rating'),
                                                       F.count('rating').alias('n_user_ratings'),
                                                       F.countDistinct('id').alias('n_recipes')))

interactions, recipes  =  interaction_level_df.count(), interaction_level_df.agg(F.countDistinct('id')).first()[0]

tags_ratings_summary = (tags_ratings_summary.withColumn("in_percent_recipies", F.col ("n_recipes")/F.lit(recipes))
                                            .withColumn("in_percent_interactions", F.col ("n_user_ratings")/F.lit(interactions)))

#### 1. Top ```n``` most rated tags

tags_ratings_summary.sort(col('n_user_ratings').desc()).show(20)

Drop tags present in majority of recipes more than 75% rows 

tags_ratings_summary_raw = tags_ratings_summary

tags_ratings_summary = tags_ratings_summary_raw.filter('in_percent_interactions<=0.75')

# print the quantiles of in_percent_interactions  
get_quantiles(df = tags_ratings_summary , 
              col_name = 'in_percent_interactions' , 
              quantiles_list = [0.01,0.25,0.5, 0.75,0.8,0.85,0.9,0.95, 0.99])

# keeping the tags appearing in the top 5 percentile 
top_most_frequent_tags = tags_ratings_summary.filter('in_percent_interactions>=0.16')

top_most_frequent_tags.count()

top_frequent_tags_list = list(top_most_frequent_tags.toPandas()['individual_tag'])

interaction_level_df = add_OHE_columns(interaction_level_df,top_frequent_tags_list)

#### 2.  Bottom ```n``` least rated tags

tags_ratings_summary.sort(col('in_percent_interactions')).show(5)

The above tags are present in 1 recipe in over two hundred thousand. The features we create based on these tags will not teach the model new information. If these tags were one hot encoded, the entire column would be filled with zeros, and only a few rows will have 1s. One hot encoding of these tags is not a good idea. If you come up with an encoding that captures the rarity of these tags, only then can you add these tags to the analysis.

#### 3. Top ```n``` rated tags 

tags_ratings_summary.sort(col('avg_user_rating').desc()).show(5)

Top rated tags have low number of ratings.

get_quantiles (df = tags_ratings_summary, 
               col_name ='n_user_ratings', 
               quantiles_list = [0.01, 0.05, 0.1, 0.15, 0.2, 0.25, 0.5, 0.75, 0.99])

tags_ratings_summary_raw2 = tags_ratings_summary

# Dropping the tags with less than 100 ratings 
tags_ratings_summary = tags_ratings_summary.filter('n_user_ratings >=100')

get_quantiles(df = tags_ratings_summary, 
              col_name = 'avg_user_rating',
              quantiles_list = [0.01,0.25,0.5, 0.75,0.8,0.85,0.9,0.95, 0.99])

# keeping tags above 95 percentile
top_rated_tags_df = tags_ratings_summary.filter('avg_user_rating>=4.53')

top_rated_tags_df.count()

top_rated_tags_list = list(top_rated_tags_df.toPandas()['individual_tag'])

Check if any of the current tags have been added earlier

set(top_frequent_tags_list) & set(top_rated_tags_list)

all_added_columns_set = set(top_frequent_tags_list).union(set(top_rated_tags_list))

interaction_level_df = add_OHE_columns(interaction_level_df,top_rated_tags_list)

#### 3. Bottom ```n``` rated tags 

get_quantiles (df = tags_ratings_summary, 
               col_name ='n_user_ratings', 
               quantiles_list = [0.01, 0.05, 0.1, 0.15, 0.2, 0.25, 0.5, 0.75, 0.99])

# checking for all cases within bottom 1 percentile
tags_ratings_summary.filter('avg_user_rating<=4').count()

bottom_rated_tags_list = list(tags_ratings_summary.filter('avg_user_rating<=4').toPandas()['individual_tag'])
bottom_rated_tags_list

Check if any of the current tags have been added earlier

all_added_columns_set & set(bottom_rated_tags_list)

interaction_level_df = add_OHE_columns(interaction_level_df,bottom_rated_tags_list)

## Final DataFrame

len(interaction_level_df.columns)

interaction_level_df.write.mode('overwrite').parquet('s3://recass/interaction_level_df_post_part2_feature_extraction.parquet')
