==> 04_Train_Test_Split.ipynb <==
## Setup

import os
# give googe drive the required permission
from google.colab import drive
drive.mount('/content/drive')


os.chdir("/content/drive/MyDrive/food_recommender_assignment/")
os.getcwd()

os.chdir("/content/drive/MyDrive/food_recommender_assignment/Dataset/")
os.getcwd()

try:
  import pyspark 
except:
  !pip install pyspark==3.1.2
  import pyspark

from pyspark.sql import SparkSession
from pyspark.context import SparkContext

spark = SparkSession.builder.master("local").config('spark.ui.port', '4050').getOrCreate()

spark

## Imports

import pandas as pd 
import numpy as np

from pyspark.sql import functions as F
# Import for typecasting columns
from pyspark.sql.types import IntegerType,BooleanType,DateType,FloatType,StringType, ArrayType

## Read the data 

raw_ratings_df = (spark.read.csv("raw_ratings_small.csv", 
                                 header=True, 
                                 inferSchema= True))

raw_recipes_df = spark.read.csv("raw_recipies_small.csv", 
                                header=True, 
                                inferSchema=True)

#### Decide a split date based on the ratings dataframe. 

# Finding the number of data points in the interaction dataset. 
num_review_int = raw_ratings_df.count()

num_review_int

Divide the data into train and test based on the 80 - 20 split using the approach discussed.

test_num_reviews_int = round(num_review_int *0.2)

# Sorting the interactions dataset in descending order of review date. 
# Extracting ```test_num_reviews_int``` most recent reviews. 

temp_ratings_df = (raw_ratings_df.sort("review_date", ascending=False)
                                 .limit(test_num_reviews_int)
                  )

temp_ratings_df.collect()[-1][4]

temp_ratings_df.count()

Split the data into two parts before and after 2011-07-17. 

- All reviews in the ratings data after 2011-07-17 will not exsist in the training set. 
- For all future predictions the date will be set at 2011-07-18.   

# Join raw_recipes and raw_ratings
# Using recipe_id as the key to join these dataframes 
# The resulting dataframe must have all rows from the raw_ratings dataframe. 

interaction_level_df = raw_ratings_df.join(raw_recipes_df,raw_ratings_df.recipe_id == raw_recipes_df.id,'right' )

interaction_level_df.printSchema()

interaction_level_df.show(5)

interaction_level_df = (interaction_level_df
                        .withColumn('review_date', F.to_date(interaction_level_df['review_date']) 
                                   ))
interaction_level_df = (interaction_level_df
                        .withColumn('submitted', F.to_date(interaction_level_df['submitted']) 
                                   ))

interaction_level_df.filter( F.col('review_date') >= F.to_date(F.lit('2011-07-17'))).count()

len(interaction_level_df.columns)

interaction_level_df.filter( "review_date >= '2011-07-17'").orderBy(F.to_date(F.col('review_date')).asc()).show(5)

# Using the filter command to separate the datasets. 
# All interactions which were rated BEFORE '2011-07-17' will be train data. 

train_interaction_level_df  = (interaction_level_df.filter(  "review_date <'2011-07-17'"  ))

# Using the filter command to separate the datasets. 
# All interactions which were rated AFTER '2011-07-17' will be train data. 

test_interaction_level_all_recipies_df  = (interaction_level_df.filter(   "review_date >= '2011-07-17'"))

# creating data files for modeling 

(train_interaction_level_df.coalesce(1)
                           .write.mode('overwrite')
                           .parquet('train_interaction_level_df.parquet'))  # change the file name and file path

(test_interaction_level_all_recipies_df.coalesce(1)
                                       .write.mode('overwrite')
                                       .parquet('test_interaction_level_df.parquet'))  # change the file name and file path

os.getcwd()
