==> 05_ALS_Algorithm.ipynb <==
## Setup

import os
# give googe drive the required permission
from google.colab import drive
drive.mount('/content/drive')

os.chdir("/content/drive/MyDrive/food_recommender_assignment/")
os.getcwd()

try:
  import pyspark 
except:
  !pip install pyspark==3.1.2
  import pyspark

try:
  import lenskit 
except:
  %pip install lenskit

from pyspark.sql import SparkSession
from pyspark.context import SparkContext

spark = SparkSession.builder.master("local").config('spark.ui.port', '4050').getOrCreate()

spark

## Imports

# import necessary libraries 
import pandas as pd
import numpy as np

from pyspark.sql import functions as F
from pyspark.sql.types import IntegerType,BooleanType,DateType,FloatType,StringType, ArrayType

# Import the required functions for ALS and estimating
from pyspark.ml.evaluation import RegressionEvaluator
from pyspark.ml.recommendation import ALS

from lenskit import batch, topn, util
from lenskit.algorithms import Recommender, als, item_knn as knn

## Read the data

To connect the data files with your google collab, download them and upload them to your home folder in your google drive. 

train_ratings_df = spark.read.parquet('Dataset/train_interaction_level_df.parquet', 
                                      header=True, 
                                      inferSchema=True)

test_ratings_all_df  = spark.read.parquet('Dataset/test_interaction_level_df.parquet', 
                                          header=True, 
                                          inferSchema=True)

raw_recipes_df = spark.read.csv("RAW_recipes_cleaned.csv", 
                                header=True, 
                                inferSchema=True)

train_ratings_df.select(F.countDistinct("user_id")).collect()[0][0]

train_ratings_df.select("rating").count()

numerator = train_ratings_df.select("rating").count()


num_users = train_ratings_df.select(F.countDistinct("user_id")).collect()[0][0]
num_recipes = train_ratings_df.select(F.countDistinct("recipe_id")).collect()[0][0]

denominator = num_users * num_recipes

sparsity = (1.0 - (numerator *1.0)/denominator)*100
print("The training dataframe is ", "%.7f" % sparsity + "% empty.")

## Functions

def manual_recommendation_check (user_id):
  '''
  Given a user ID form the test dataset, this function will return the names of the recipes recommended to the user. 

  Initialize this function after the all_recs data frame is calculated. 

  Input user_id of a user from the test set as an integer.  
  Prints the names of recipes recommended to this user. 
  Returns nothing. 
  '''
  recs_user = all_recs_als[all_recs_als.user == user_id]
  recs_user_list = list(recs_user.item.values)
  recs_user_list = [x.item() for x in recs_user_list]
  display((raw_recipes_df.filter(F.col('id').isin(recs_user_list))
               .select("name")
               .collect()
               ))

## Model

#### Task 02 - Collaborative Filtering Model

Add the argument details in the algorithm initialization below to build the ALS model.

# Create ALS model
als = ALS(userCol= 'user_id', #  name of the column for users  
          itemCol= 'recipe_id', # name of the column for recipes
          ratingCol= 'rating'  , # name of the column for ratings 
          nonnegative = True, 
          implicitPrefs = False, 
          coldStartStrategy="drop"
         )

assert type(als) == pyspark.ml.recommendation.ALS

#### Training

#Fit the model to the 'train' dataset
model = als.fit(train_ratings_df)

## Prediction 

test_ratings_all_df.count()

# use the model to create predictions for test data
test_predictions_unseen = model.transform(test_ratings_all_df) 

test_predictions_unseen.count()

test_predictions_unseen.select(F.col("user_id"),
                               F.col("recipe_id"),
                               F.col("rating"),
                               F.col("prediction")
                              ).show(5)

# Each user in the test set must have 10 predictions. 
# Use ALS model to get these predictions. 
# You can use the recommendForAllUsers() method. 

recommendations = model.recommendForAllUsers(10) 

# Transfroming the results by exploding the recommendations column

recommendations = (recommendations.select(F.col("user_id"),
                                          F.posexplode(F.col("recommendations")).alias("pos", "item")) 
                                  .select(F.col("user_id"),
                                          F.col("pos"), 
                                          F.col("item.recipe_id").alias("recomended_recipe_id"), 
                                          F.col("item.rating").alias("predicted_rating")))

## Evaluation

### Task 04 - Model Evaluation

#### 1. RMSE

Add the argument details to the evaluator function below to calculate the RMSE score of the ALS model. 

# Define evaluator as RMSE and print RMSE value
evaluator_seen = RegressionEvaluator(metricName="rmse", 
                                     labelCol= 'rating',  # name of the column that has the ratings 
                                     predictionCol= 'prediction'# name of the column that has the predicted ratings. 
                                     ) 

RMSE = evaluator_seen.evaluate(test_predictions_unseen ) #  evaluator to find the RMSE on the test set. 
print(RMSE)

RMSE = evaluator_seen.evaluate( ) #  evaluator to find the RMSE on the test set. 
print(RMSE)

#### Rank based Metrics

We will use the lenskit library to calculate the ranking-based matrics. The lenskit library is available in pandas only so we need to convert the data frames from PySpark dataframes to Pandas dataframes. 

all_recs_als = recommendations.toPandas()

all_recs_als


column_names = {'user_id':'user',	'pos':'rank',	'recomended_recipe_id':'item',	
                'predicted_rating':'score'} # a dictionary with current column names as keys and the intended column names as values. 
all_recs_als = all_recs_als.rename(columns=column_names)

all_recs_als

# Adding a column to make sure the we know these recommendations are from the ALS algorithm. 

all_recs_als["algorithm"] = "ALS"

test_ratings_all_df.select(F.col('user_id').alias('user'),
                               F.col('recipe_id').alias('item'),
                               F.col('rating')).show(2)

test_data = (test_ratings_all_df.select(F.col('user_id').alias('user'),
                               F.col('recipe_id').alias('item'),
                               F.col('rating')).toPandas())

test_data

# code to calculate the necessary metrics 
# the code below uses lenskit library to evaluate the rank metrics

rla = topn.RecListAnalysis()
rla.add_metric(topn.recip_rank)
rla.add_metric(topn.ndcg)
rla.add_metric(topn.dcg)
results = rla.compute(all_recs_als, test_data)
results.head()

## Manual Prediction Checking

1. Why are ndcg and dcg nulls? EG: user 28170

test_data[test_data.user == 28170]

all_recs_als[all_recs_als.user == 28170]

User ```28170``` does not appear in the test set. Hence cannot be evaluated. 

2. Why are all metrics 0 for specific users?

test_data[test_data.user == 56680]

all_recs_als[all_recs_als.user == 56680]

User-recipe combination does not appear in the recommendations set. 

3. Are any non zero metrics? 

results[results.ndcg > 0]

results[results.dcg > 0]

results[results.recip_rank > 0]

There are few user recipe combination that has a corresponding value in the test set. Hence, we have only few non-zero value of evaluation metrics.


Manually, check one prediction. 

manual_recommendation_check(653438)

Of the recipes that have been recommended, few appear similar, and few do not. 

## Saving the models and predictions

all_recs_als.to_csv("recommendation_als.csv", 
                    index=False)

model.save('ALS_model.model')
